finished overhaul of list, array and vector

+++ vector.h
@@ -10,12 +10,12 @@ enum {
 	VEC_AXIS_Y,
 	VEC_AXIS_Z,
 	VEC_AXIS_LAST,
-};
+};  
 extern size_t VEC_AXIS[VEC_AXIS_LAST];
 
 typedef struct _VectorIF {
-	double (*get)(void* obj, unsigned int index);
-	void (*set)(void* obj, unsigned int index, double value);
+	long double (*get)(void* obj, size_t index);
+	void (*set)(void* obj, size_t index, double value);
 	double (*calcDotP)(void* vec1, void* vec2);
 	void (*rotate3D)(void* obj, double angle, size_t axis); //3D only
 	void (*rotate2D)(void* obj, double angle); //2D only
